CR changes, protocol implementation, deleted NewMessage
